image-ocr/pyproject.toml
[project]
name = "image-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "easyocr>=1.7.2",
    "matplotlib>=3.10.3",
    "openai>=1.79.0",
    "python-dotenv>=1.1.0",
]
